restoring deployment directory to it's original contents
