+++ service/worker/deployment/deployment_util.go
@@ -40,10 +40,6 @@ import (
 
 const (
 	// Workflow types
-	WorkerDeploymentWorkflowType        = "temporal-sys-worker-deployment-workflow"
-	WorkerDeploymentVersionWorkflowType = "temporal-sys-worker-deployment-version-workflow"
-
-	// [cleanup-wv]
 	DeploymentWorkflowType       = "temporal-sys-deployment-workflow"
 	DeploymentSeriesWorkflowType = "temporal-sys-deployment-series-workflow"
